Add combo as objective

+++ Assets/_MainAssets/Feature/Combo/ComboSystem.cs
@@ -72,7 +72,7 @@ namespace Combo
             _timer.Resume();
         }
 
-        public float GetComboCount()
+        public int GetComboCount()
         {
             return _comboCount;
         }
